services/nanovlm/requirements.txt
torch
torchvision
transformers
accelerate
pillow
fastapi
uvicorn
python-multipart
peft
bitsandbytes
